+++ agrest-engine/src/main/java/io/agrest/runtime/protocol/IUpdateRequestParser.java
@@ -4,14 +4,14 @@ import io.agrest.EntityUpdate;
 import io.agrest.meta.AgEntity;
 
 import java.io.InputStream;
-import java.util.Collection;
+import java.util.List;
 
 /**
  * @since 5.0
  */
 public interface IUpdateRequestParser {
 
-    <T> Collection<EntityUpdate<T>> parse(AgEntity<T> entity, String entityData);
+    <T> List<EntityUpdate<T>> parse(AgEntity<T> entity, String entityData);
 
-    <T> Collection<EntityUpdate<T>> parse(AgEntity<T> entity, InputStream entityStream);
+    <T> List<EntityUpdate<T>> parse(AgEntity<T> entity, InputStream entityStream);
 }